chore(package): split out proptest/arbitrary details to new module
